Address Copilot review: shared HWM, walkJsonl helper, late session_meta
- ingestAll loads HWM once and saves once, removing double I/O and reducing
  clobber risk when multiple passes share state.
- Factor walkJsonl into packages/cli/src/walk.ts; reuse from ingest and codex
  wrapper to keep traversal logic in one place.
- Drop the mtime-vs-spawn-start filter in findNewSessions; preSnapshot already
  excludes pre-existing files, and the 1ms cushion is fragile on filesystems
  with coarse timestamp granularity.
- If session_meta arrives after task_started, propagate its cwd to the open
  turn's project when still unset.
