+++ packages/cli/src/commands/codex.ts
@@ -1,6 +1,4 @@
 import { spawn } from 'node:child_process';
-import { readdir, stat } from 'node:fs/promises';
-import type { Dirent } from 'node:fs';
 import { homedir } from 'node:os';
 import * as path from 'node:path';
 
@@ -9,6 +7,7 @@ import { appendTurns, stamp } from '@relayburn/ledger';
 import type { Enrichment } from '@relayburn/ledger';
 
 import type { ParsedArgs } from '../args.js';
+import { walkJsonl } from '../walk.js';
 
 const CODEX_SESSIONS = path.join(homedir(), '.codex', 'sessions');
 
@@ -31,7 +30,7 @@ export async function runCodexWrapper(args: ParsedArgs): Promise<number> {
     });
   });
 
-  const newFiles = await findNewSessions(preSnapshot, spawnStartTs);
+  const newFiles = await findNewSessions(preSnapshot);
   if (newFiles.length === 0) {
     process.stderr.write(`[burn] no new codex session files found under ${CODEX_SESSIONS}\n`);
     return code;
@@ -55,38 +54,7 @@ async function snapshotSessions(): Promise<Set<string>> {
   return out;
 }
 
-async function findNewSessions(pre: Set<string>, spawnStartTs: number): Promise<string[]> {
+async function findNewSessions(pre: Set<string>): Promise<string[]> {
   const now = await walkJsonl(CODEX_SESSIONS);
-  const candidates: string[] = [];
-  for (const file of now) {
-    if (pre.has(file)) continue;
-    try {
-      const st = await stat(file);
-      if (st.mtimeMs + 1 < spawnStartTs) continue;
-      candidates.push(file);
-    } catch {
-      continue;
-    }
-  }
-  return candidates;
-}
-
-async function walkJsonl(root: string): Promise<string[]> {
-  const out: string[] = [];
-  const stack: string[] = [root];
-  while (stack.length > 0) {
-    const dir = stack.pop()!;
-    let entries: Dirent[];
-    try {
-      entries = (await readdir(dir, { withFileTypes: true })) as Dirent[];
-    } catch {
-      continue;
-    }
-    for (const e of entries) {
-      const full = path.join(dir, e.name);
-      if (e.isDirectory()) stack.push(full);
-      else if (e.isFile() && e.name.endsWith('.jsonl')) out.push(full);
-    }
-  }
-  return out;
+  return now.filter((file) => !pre.has(file));
 }
