+++ packages/cli/src/ingest.ts
@@ -1,10 +1,11 @@
 import { readdir, stat } from 'node:fs/promises';
-import type { Dirent } from 'node:fs';
 import { homedir } from 'node:os';
 import * as path from 'node:path';
 
 import { parseClaudeSession, parseCodexSession, type TurnRecord } from '@relayburn/reader';
 import { appendTurns, loadHwm, saveHwm, type HwmMap } from '@relayburn/ledger';
+
+import { walkJsonl } from './walk.js';
 
 const CLAUDE_PROJECTS = path.join(homedir(), '.claude', 'projects');
 const CODEX_SESSIONS = path.join(homedir(), '.codex', 'sessions');
@@ -17,8 +18,30 @@ export interface IngestReport {
 
 export async function ingestClaudeProjects(): Promise<IngestReport> {
   const hwm = await loadHwm();
-  const report = { scannedSessions: 0, ingestedSessions: 0, appendedTurns: 0 };
+  const report = emptyReport();
+  await ingestClaudeInto(hwm, report);
+  await saveHwm(hwm);
+  return report;
+}
 
+export async function ingestCodexSessions(): Promise<IngestReport> {
+  const hwm = await loadHwm();
+  const report = emptyReport();
+  await ingestCodexInto(hwm, report);
+  await saveHwm(hwm);
+  return report;
+}
+
+export async function ingestAll(): Promise<IngestReport> {
+  const hwm = await loadHwm();
+  const report = emptyReport();
+  await ingestClaudeInto(hwm, report);
+  await ingestCodexInto(hwm, report);
+  await saveHwm(hwm);
+  return report;
+}
+
+async function ingestClaudeInto(hwm: HwmMap, report: IngestReport): Promise<void> {
   const projects = await listDirs(CLAUDE_PROJECTS);
   for (const projectDir of projects) {
     const files = await listJsonlFiles(projectDir);
@@ -26,31 +49,16 @@ export async function ingestClaudeProjects(): Promise<IngestReport> {
       await ingestOne(file, hwm, report, (f) => parseClaudeSession(f, { sessionPath: f }));
     }
   }
-
-  await saveHwm(hwm);
-  return report;
 }
 
-export async function ingestCodexSessions(): Promise<IngestReport> {
-  const hwm = await loadHwm();
-  const report = { scannedSessions: 0, ingestedSessions: 0, appendedTurns: 0 };
-
+async function ingestCodexInto(hwm: HwmMap, report: IngestReport): Promise<void> {
   for (const file of await walkJsonl(CODEX_SESSIONS)) {
     await ingestOne(file, hwm, report, (f) => parseCodexSession(f, { sessionPath: f }));
   }
-
-  await saveHwm(hwm);
-  return report;
 }
 
-export async function ingestAll(): Promise<IngestReport> {
-  const a = await ingestClaudeProjects();
-  const b = await ingestCodexSessions();
-  return {
-    scannedSessions: a.scannedSessions + b.scannedSessions,
-    ingestedSessions: a.ingestedSessions + b.ingestedSessions,
-    appendedTurns: a.appendedTurns + b.appendedTurns,
-  };
+function emptyReport(): IngestReport {
+  return { scannedSessions: 0, ingestedSessions: 0, appendedTurns: 0 };
 }
 
 async function ingestOne(
@@ -85,26 +93,6 @@ async function ingestOne(
     lastTs: last.ts,
     mtimeMs: st.mtimeMs,
   };
-}
-
-async function walkJsonl(root: string): Promise<string[]> {
-  const out: string[] = [];
-  const stack: string[] = [root];
-  while (stack.length > 0) {
-    const dir = stack.pop()!;
-    let entries: Dirent[];
-    try {
-      entries = (await readdir(dir, { withFileTypes: true })) as Dirent[];
-    } catch {
-      continue;
-    }
-    for (const e of entries) {
-      const full = path.join(dir, e.name);
-      if (e.isDirectory()) stack.push(full);
-      else if (e.isFile() && e.name.endsWith('.jsonl')) out.push(full);
-    }
-  }
-  return out;
 }
 
 async function listDirs(parent: string): Promise<string[]> {
