+++ packages/cli/src/walk.ts
@@ -1,0 +1,23 @@
+import { readdir } from 'node:fs/promises';
+import type { Dirent } from 'node:fs';
+import * as path from 'node:path';
+
+export async function walkJsonl(root: string): Promise<string[]> {
+  const out: string[] = [];
+  const stack: string[] = [root];
+  while (stack.length > 0) {
+    const dir = stack.pop()!;
+    let entries: Dirent[];
+    try {
+      entries = (await readdir(dir, { withFileTypes: true })) as Dirent[];
+    } catch {
+      continue;
+    }
+    for (const e of entries) {
+      const full = path.join(dir, e.name);
+      if (e.isDirectory()) stack.push(full);
+      else if (e.isFile() && e.name.endsWith('.jsonl')) out.push(full);
+    }
+  }
+  return out;
+}
